Finish basic mapstat testing.
Could do bad actor testing but so far, the test are proving accurate.

Updated database scripts to include default values for t1/t2 scores.

Added todo in `util.js` to check for API key before authenticating.

+++ routes/mapstats.js
@@ -17,9 +17,8 @@ const router = express.Router();
 const db = require("../db");
 
 /** Utility class for various methods used throughout.
-* @const */
-const Utils = require('../utility/utils');
-
+ * @const */
+const Utils = require("../utility/utils");
 
 /** GET - Route serving to get all game servers.
  * @name router.get('/')
@@ -34,8 +33,8 @@ router.get("/", async (req, res, next) => {
     // Check if admin, if they are use this query.
     let sql = "SELECT * FROM map_stats";
     const allStats = await db.query(sql);
-    if (allStats.length === 0){
-      res.status(404).json({message: "No stats found."});
+    if (allStats.length === 0) {
+      res.status(404).json({ message: "No stats found." });
       return;
     }
     res.json(allStats);
@@ -57,8 +56,8 @@ router.get("/:match_id", async (req, res, next) => {
     matchID = req.params.match_id;
     let sql = "SELECT * FROM map_stats where match_id = ?";
     const mapStats = await db.query(sql, matchID);
-    if (mapStats.length === 0){
-      res.status(404).json({message: "No stats found."});
+    if (mapStats.length === 0) {
+      res.status(404).json({ message: "No stats found." });
       return;
     }
     res.json(mapStats);
@@ -76,28 +75,49 @@ router.get("/:match_id", async (req, res, next) => {
  * @param {string} req.body[0].map_name - The current map name.
  * @param {DateTime} req.body[0].start_time - The start time, as DateTime.
  *
-*/
+ */
 router.post("/create", Utils.ensureAuthenticated, async (req, res, next) => {
-  try{
-    // Get the match and see if the user owns said match, or is an admin.
-    await db.withTransaction(db, async () => {
-      let userProfile = req.user.id;
-      let matchId = req.body[0].match_id;
-      let mapNum = req.body[0].map_number;
-      let mapName = req.body[0].map_name;
-      let startTime =  req.body[0].start_time;
-      let matchQuery = "SELECT * FROM `match` WHERE id = ?"
-      let sql = "INSERT INTO map_stats (match_id, map_number, map_name, start_time) VALUES (?,?,?,?)";
-      const matchRecord = await db.query(matchQuery, [matchId]);
-      if(req.user.super_admin === 1 || req.user.admin === 1 || userProfile === matchRecord[0].user_id){
-        await db.query(sql, [matchId, mapNum, mapName, startTime]);
-        res.json("Map stats inserted successfully!");
-      } else {
-        res.status(401).json({message: "You are not authorized to perform this action."});
-      }
-    });
-  } catch ( err ) {
-    res.status(500).json({ message: err.toString() })
+  try {
+    if (req.body[0].match_id == null) {
+      res.status(404).json({ message: "Match ID Not Provided" });
+      return;
+    }
+    let currentMatchInfo =
+      "SELECT user_id, cancelled, forfeit, end_time FROM `match` WHERE id = ?";
+    const matchRow = await db.query(currentMatchInfo, req.body[0].match_id);
+    if (matchRow.length === 0) {
+      res.status(404).json({ message: "No match found." });
+      return;
+    } else if (
+      matchRow[0].user_id != req.user.id &&
+      !Utils.superAdminCheck(req.user)
+    ) {
+      res
+        .status(401)
+        .json({ message: "User is not authorized to perform action." });
+      return;
+    } else if (
+      matchRow[0].cancelled == 1 ||
+      matchRow[0].forfeit == 1 ||
+      matchRow[0].end_time != null
+    ) {
+      res.status(401).json({ message: "Match is already finished." });
+      return;
+    } else {
+      await db.withTransaction(db, async () => {
+        let mapStatSet = {
+          match_id: req.body[0].match_id,
+          map_number: req.body[0].map_number,
+          map_name: req.body[0].map_name,
+          start_time: req.body[0].start_time,
+        };
+        let sql = "INSERT INTO map_stats SET ?";
+        await db.query(sql, [mapStatSet]);
+        res.json({ message: "Map stats inserted successfully!" });
+      });
+    }
+  } catch (err) {
+    res.status(500).json({ message: err.toString() });
   }
 });
 
@@ -113,40 +133,66 @@ router.post("/create", Utils.ensureAuthenticated, async (req, res, next) => {
  * @param {string} [req.body[0].map_name] - The map the stats are recording from in the match series.
  * @param {string} [req.body[0].demo_file] - The demo file of the match once demo has been finished recording.
  *
-*/
+ */
 router.put("/update", Utils.ensureAuthenticated, async (req, res, next) => {
-  try{
-    await db.withTransaction(db, async () => {
-      let mapStatId = req.body[0].map_stats_id;
-      let userProfile = req.user.id;
-      let matchQuery = "SELECT a.user_id FROM `match` a, map_stats b WHERE b.id = ?";
-      let updatedValues = {
-        end_time: req.body[0].end_time,
-        team1_score: req.body[0].team1_score,
-        team2_score: req.body[0].team2_score,
-        winner: req.body[0].winner,
-        demoFile: req.body[0].demo_file,
-        map_name: req.body[0].map_name
-      };
-      updatedvalues = await db.buildUpdateStatement(updatedValues);
-      if(Object.keys(updatedvalues).length === 0){
-        res.status(412).json({message: "No update data has been provided."});
-        return;
-      }
-      let sql = "UPDATE map_stats SET ? WHERE id = ?";
-
-      const matchOwner = await db.query(matchQuery, [mapStatId]);
-      if(req.user.super_admin === 1 || req.user.admin === 1 || userProfile === matchOwner[0].user_id){
+  try {
+    if (req.body[0].map_stats_id == null) {
+      res.status(404).json({ message: "Match ID Not Provided" });
+      return;
+    }
+    let currentMatchInfo =
+      "SELECT mtch.user_id as user_id, mtch.cancelled as cancelled, mtch.forfeit as forfeit, mtch.end_time as mtch_end_time FROM `match` mtch, map_stats mstat WHERE mstat.id = ? AND mstat.match_id = mtch.id";
+    const matchRow = await db.query(currentMatchInfo, req.body[0].map_stats_id);
+    if (matchRow.length === 0) {
+      res.status(404).json({ message: "No match found." });
+      return;
+    } else if (
+      matchRow[0].user_id != req.user.id &&
+      !Utils.superAdminCheck(req.user)
+    ) {
+      res
+        .status(401)
+        .json({ message: "User is not authorized to perform action." });
+      return;
+    } else if (
+      matchRow[0].cancelled == 1 ||
+      matchRow[0].forfeit == 1 ||
+      matchRow[0].mtch_end_time != null
+    ) {
+      res.status(401).json({ message: "Match is already finished." });
+      return;
+    } else {
+      await db.withTransaction(db, async () => {
+        let mapStatId = req.body[0].map_stats_id;
+        let userProfile = req.user.id;
+        let matchQuery =
+          "SELECT a.user_id FROM `match` a, map_stats b WHERE b.id = ?";
+        let updatedValues = {
+          end_time: req.body[0].end_time,
+          team1_score: req.body[0].team1_score,
+          team2_score: req.body[0].team2_score,
+          winner: req.body[0].winner,
+          demoFile: req.body[0].demo_file,
+          map_name: req.body[0].map_name,
+        };
+        updatedvalues = await db.buildUpdateStatement(updatedValues);
+        if (Object.keys(updatedvalues).length === 0) {
+          res
+            .status(412)
+            .json({ message: "No update data has been provided." });
+          return;
+        }
+        let sql = "UPDATE map_stats SET ? WHERE id = ?";
         updateMapStats = await db.query(sql, [updatedValues, mapStatId]);
         if (updateMapStats.affectedRows > 0)
-          res.json("Map Stats updated successfully!");
+          res.json({ message: "Map Stats updated successfully!" });
         else
-          res.status(401).json({message: "ERROR - Maps Stats not updated or found."});
-      } else {
-
-      }
-    });
-  } catch ( err ) {
+          res
+            .status(401)
+            .json({ message: "ERROR - Maps Stats not updated or found." });
+      });
+    }
+  } catch (err) {
     console.log(err);
     res.status(500).json({ message: err.toString() });
   }
@@ -158,30 +204,52 @@ router.put("/update", Utils.ensureAuthenticated, async (req, res, next) => {
  * @function
  * @param {int} req.body[0].map_stats_id - The ID of the map stats being removed.
  *
-*/
-router.delete("/delete", Utils.ensureAuthenticated, async (req,res,next) => {
+ */
+router.delete("/delete", Utils.ensureAuthenticated, async (req, res, next) => {
   try {
-    await db.withTransaction (db, async () => {
-      let userProfile = req.user.id; // Brought in from steam passport.
-      let mapStatsId = req.body[0].map_stats_id;
-      let deleteSql = "DELETE FROM map_stats WHERE id = ?"
-      let ownerSql = "SELECT a.user_id FROM `match` a, map_stats b WHERE b.id = ?";
-      const matchOwner = await db.query(ownerSql, [mapStatId]);
-      if(req.user.super_admin === 1 || req.user.admin === 1 || userProfile === matchOwner[0].user_id){
+    if (req.body[0].map_stats_id == null) {
+      res.status(404).json({ message: "Map Stats ID Not Provided" });
+      return;
+    }
+    let currentMatchInfo =
+      "SELECT mtch.user_id as user_id, mtch.cancelled as cancelled, mtch.forfeit as forfeit, mtch.end_time as mtch_end_time FROM `match` mtch, map_stats mstat WHERE mstat.id = ? AND mstat.match_id = mtch.id";
+    const matchRow = await db.query(currentMatchInfo, req.body[0].map_stats_id);
+    if (matchRow.length === 0) {
+      res.status(404).json({ message: "No match found." });
+      return;
+    } else if (
+      matchRow[0].user_id != req.user.id &&
+      !Utils.superAdminCheck(req.user)
+    ) {
+      res
+        .status(401)
+        .json({ message: "User is not authorized to perform action." });
+      return;
+    } else if (
+      matchRow[0].cancelled == 1 ||
+      matchRow[0].forfeit == 1 ||
+      matchRow[0].mtch_end_time != null
+    ) {
+      res.status(401).json({ message: "Match is already finished." });
+      return;
+    } else {
+      await db.withTransaction(db, async () => {
+        let userProfile = req.user.id; // Brought in from steam passport.
+        let mapStatsId = req.body[0].map_stats_id;
+        let deleteSql = "DELETE FROM map_stats WHERE id = ?";
         const delRows = await db.query(deleteSql, [mapStatsId]);
         if (delRows.affectedRows > 0)
-          res.json("Map Stats deleted successfully!");
+          res.json({ message: "Map Stats deleted successfully!" });
         else
-          res.status(401).json("ERR - Unauthorized to delete OR not found.");
-      } else {
-        res.status(401).json("You are not authorized to perform this action.");
-      }
-    });
-  } catch( err ){
+          res
+            .status(401)
+            .json({ message: "ERR - Unauthorized to delete OR not found." });
+      });
+    }
+  } catch (err) {
     console.log(err);
-    res.statuss(500).json({message: err});
+    res.status(500).json({ message: err });
   }
 });
-
 
 module.exports = router;
